requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
pymongo==3.12.3
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
tqdm==4.63.0
urllib3==1.26.8
